// types/model.ts
export interface Collection {
  primary_asset_contracts: any[]
  traits: {}
  banner_image_url: string
  chat_url?: any
  created_date: Date
  default_to_fiat: boolean
  description: string
  dev_buyer_fee_basis_points: string
  dev_seller_fee_basis_points: string
  discord_url?: any
  external_url?: any
  featured: boolean
  featured_image_url?: any
  hidden: boolean
  safelist_request_status: string
  image_url: string
  is_subject_to_whitelist: boolean
  large_image_url?: any
  medium_username?: any
  name: string
  only_proxied_transfers: boolean
  opensea_buyer_fee_basis_points: string
  opensea_seller_fee_basis_points: string
  payout_address: string
  require_email: boolean
  short_description?: any
  slug: string
  telegram_url?: any
  twitter_username: string
  instagram_username?: any
  wiki_url?: any
  ultcube_featured?: boolean
}